Switch to using injectable to setup route for serving static files
