+++ src/main/router/router.injectable.ts
@@ -4,8 +4,6 @@
  */
 import { getInjectable, getInjectionToken } from "@ogre-tools/injectable";
 import { Router, Route } from "../router";
-import isDevelopmentInjectable from "../../common/vars/is-development.injectable";
-import httpProxy from "http-proxy";
 
 export const routeInjectionToken = getInjectionToken<Route>({
   id: "route-injection-token",
@@ -15,12 +13,7 @@ const routerInjectable = getInjectable({
   id: "router",
 
   instantiate: (di) => {
-    const isDevelopment = di.inject(isDevelopmentInjectable);
-    const proxy = isDevelopment ? httpProxy.createProxy() : undefined;
-
-    const router = new Router({
-      httpProxy: proxy,
-    });
+    const router = new Router();
 
     const routes = di.injectMany(routeInjectionToken);
